requirements.txt
streamlit
python-dotenv
google-cloud-aiplatform  # for genai library (assuming Google Cloud AI Platform usage)
fpdf
matplotlib